add reservation to backend config
