+++ qiskit/providers/ibmq/jupyter/dashboard/utils.py
@@ -14,35 +14,7 @@
 
 """Utility functions for the IBM Quantum Experience dashboard."""
 
-from typing import Optional
 from collections import namedtuple
-from datetime import datetime, timedelta
-
-from qiskit.providers.ibmq.ibmqbackend import IBMQBackend
-from qiskit.providers.ibmq.backendreservation import BackendReservation
-
 
 BackendWithProviders = namedtuple('BackendWithProviders', ['backend', 'providers'])
 """Named tuple used to pass a backend and its providers."""
-
-
-def get_next_reservation(
-        backend: IBMQBackend,
-        time_period_hr: int = 24
-) -> Optional[BackendReservation]:
-    """Get the next reservation within the input time period for the backend.
-
-    Args:
-        backend: Backend for which the reservation is to be returned.
-        time_period_hr: Time period to search for reservations.
-
-    Returns:
-        The next reservation for the backend.
-    """
-    reservations = backend.reservations(
-        start_datetime=datetime.now(),
-        end_datetime=datetime.now() + timedelta(hours=time_period_hr))
-    if reservations:
-        next_reservation = min(reservations, key=lambda rsvr: rsvr.start_datetime)
-        return next_reservation
-    return None
